Change user interface to command/subcommand-style
Previously, everything was specified via command line options (as
opposed to positional parameters). For example, the -command (or
--command) command line option specified which command to run and the
-action (or --action) command line option specified which subcommand
to run. In addition to being unnecessarily verbose, this approach made
it easier to make mistakes when using the utility.

Now, proper command/subcommand processing is done. This means that the
utility can inform the user at each level which commands or
subcommands are available and, for each subcommand, which mandatory
positional parameters must be supplied.