Updates to set-custom-fields command line
An option (-s or --strict) has been added. If this option is specified
on the command line, if an unrecognised custom field is specified, the
program will report an error and exit with a non-zero exit
status. If this option is not specified, a warning will be printed but
the program will not immediately fail.

The -team option has been changed to -t.

Descriptions of the command, its options and parameters have been
added.