Force full scans projects that do not exist or have not been scanned
The assumption is that the force-full-scan subcommand will be called
as part of a larger flow which will subsequently create the project,
if it does not already exist, and launch a scan. For this reason, if
the project does not exist, we return an exit code indicating that a
full scan is required.