Upgrade to Gradle version 7.0.2
This fixes the Gradle build, which was broken by the upgrade to Spring
Boot 2.5.x. A casualty of the upgrade is the maven publishing: the
mechanism used for Gradle 5.x is no longer supported. As GitHub itself
seems a better channel for publication of releases, it does not seem
urgent, or even necessary, to find a way of getting this to work under
the newer version of Gradle.